Add files via upload
“Blackjack”, is one of the most popular casino games in the world. The objective of the game is to obtain a hand adding as close to 21 as possible without going over that value, i.e “busting”. It is a casino backed game, which means the player competes against the house rather than against one another directly. The player(s’) and dealer begin by being dealt two cards face up. The game then allows a turn for each player to either obtain another card (“Hit”) to try to add more to their sum or not obtain another card and keep the card total they already have (“Stand”). 

We have two main objectives we want to test. The first is to test whether or not using a different number of decks in the game changes the percentages of getting a value of exactly 21 (we assume the players can not count cards). The second objective is to test the chances a player has of winning, losing, or tying against the dealer. For this we will test the scenario of only 1 player against the dealer and we will test this percentage with different numbers of decks that are in play. We know that the house has an advantage since the player will have to make their play before the dealer, and we want to see just how that advantage changes with different number of decks.